chore(queries): delete zero-traffic breakdown_props tree
get_breakdown_prop_values has no production callers (only its own tests)
and zero tagged queries in either region across the query_log_archive
retention window. Its exclusive dependency tree goes with it: the legacy
FunnelEventQuery, EventQuery/EnterpriseEventQuery, GroupsJoinQuery,
person-on-events v2 join SQL, and the trends SQL templates.

get_breakdown_cohort_name, the one function modern funnels still import
from breakdown_props, moves to the funnels utils module.
