+++ posthog/hogql_queries/insights/funnels/base.py
@@ -22,11 +22,10 @@ from posthog.hogql.parser import parse_expr, parse_select
 from posthog.clickhouse.materialized_columns import ColumnName
 from posthog.hogql_queries.insights.funnels.funnel_event_query import FunnelEventQuery
 from posthog.hogql_queries.insights.funnels.funnel_query_context import FunnelQueryContext
-from posthog.hogql_queries.insights.funnels.utils import funnel_window_interval_unit_to_sql
+from posthog.hogql_queries.insights.funnels.utils import funnel_window_interval_unit_to_sql, get_breakdown_cohort_name
 from posthog.hogql_queries.insights.utils.breakdowns import ALL_USERS_COHORT_ID
 from posthog.hogql_queries.insights.utils.entities import is_equal, is_superset
 from posthog.models.property.property import PropertyName
-from posthog.queries.breakdown_props import get_breakdown_cohort_name
 from posthog.queries.util import correct_result_for_sampling
 from posthog.types import FunnelEntityNode
